requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
alembic==1.13.1
cryptography==42.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic[email]==2.5.3

# HTTP requests for D-Money API
httpx==0.26.0
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Date/Time handling
python-dateutil==2.8.2

# Utilities
pydantic-settings==2.1.0
email-validator==2.1.0

# Development/Testing
pytest==7.4.4
pytest-asyncio==0.23.3
